+++ shaders/player_f.glsl
@@ -19,11 +19,11 @@ uniform sampler2D shadow;
 
 varying vec4 pointLight [10];
 uniform vec4 light_color[10];
-uniform float num_lights;
+uniform int num_lights;
 
 void main()
     {    
-    if (dot(p3d_ClipPlane[0], vpos) < 0) 
+    if (dot(p3d_ClipPlane[0], vpos) < 0.0) 
         {
         discard;
         }
